refactor(task-graph): move schema helpers from Workflow.ts to GraphSchemaUtils.ts
schemaHasTypedArrayFormat, hasVectorOutput, hasVectorLikeInput, the
TYPED_ARRAY_FORMAT_PREFIX constant, and updateBoundaryTaskSchemas (was
Workflow.updateBoundaryTaskSchemas, now a free function) are general
graph/schema utilities, not Workflow concerns. Moving them out trims
~120 lines from Workflow.ts as part of the larger decomposition.
